requirements.txt
Flask==2.0.1
Pillow==8.3.2
PyPDF2==2.0.0
moviepy==1.0.3
Werkzeug==2.0.1
